Move CddAndFeeDetails trait to primitives.
