+++ pallets/common/src/traits/identity.rs
@@ -26,12 +26,11 @@ use polymesh_primitives::identity::limits::{
     MAX_ASSETS, MAX_EXTRINSICS, MAX_PALLETS, MAX_PORTFOLIOS,
 };
 use polymesh_primitives::{
-    protocol_fee::ChargeProtocolFee, secondary_key::SecondaryKey, AuthorizationData, Balance,
-    CustomClaimTypeId, IdentityClaim, IdentityId, Permissions, Ticker,
+    protocol_fee::ChargeProtocolFee, secondary_key::SecondaryKey, traits::CddAndFeeDetails,
+    AuthorizationData, Balance, CustomClaimTypeId, IdentityClaim, IdentityId, Permissions, Ticker,
 };
 
 use crate::traits::group::GroupTrait;
-use crate::traits::transaction_payment::CddAndFeeDetails;
 
 pub type AuthorizationNonce = u64;
