+++ pallets/common/src/traits/mod.rs
@@ -33,8 +33,6 @@ pub mod identity;
 pub mod multisig;
 pub mod nft;
 pub mod portfolio;
-pub mod transaction_payment;
-pub use transaction_payment::CddAndFeeDetails;
 pub mod relayer;
 pub mod settlement;
 pub mod statistics;
