+++ pallets/protocol-fee/src/lib.rs
@@ -47,10 +47,10 @@ use frame_support::{
 use frame_system::ensure_root;
 use polymesh_common_utilities::{
     identity::Config as IdentityConfig, traits::relayer::SubsidiserTrait,
-    transaction_payment::CddAndFeeDetails,
 };
 use polymesh_primitives::{
     protocol_fee::{ChargeProtocolFee, ProtocolOp},
+    traits::CddAndFeeDetails,
     Balance, IdentityId, PosRatio, GC_DID,
 };
 use sp_runtime::{traits::Zero, Perbill};
