+++ pallets/runtime/tests/src/fee_details.rs
@@ -7,8 +7,7 @@ use frame_support::assert_noop;
 use pallet_balances as balances;
 use pallet_identity as identity;
 use pallet_multisig as multisig;
-use polymesh_common_utilities::traits::transaction_payment::CddAndFeeDetails;
-use polymesh_primitives::{Signatory, TransactionError};
+use polymesh_primitives::{traits::CddAndFeeDetails, Signatory, TransactionError};
 use polymesh_runtime_develop::runtime::{CddHandler, RuntimeCall};
 use sp_keyring::AccountKeyring;
 use sp_runtime::transaction_validity::InvalidTransaction;
