+++ primitives/src/traits.rs
@@ -14,6 +14,18 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 use crate::Balance;
+use sp_runtime::transaction_validity::InvalidTransaction;
+
+// Polymesh note: This was specifically added for Polymesh
+pub trait CddAndFeeDetails<AccountId, Call> {
+    fn get_valid_payer(
+        call: &Call,
+        caller: &AccountId,
+    ) -> Result<Option<AccountId>, InvalidTransaction>;
+    fn clear_context();
+    fn set_payer_context(payer: Option<AccountId>);
+    fn get_payer_from_context() -> Option<AccountId>;
+}
 
 /// A currency that has a block rewards reserve.
 pub trait BlockRewardsReserveCurrency<NegativeImbalance> {
